requirements.txt
# Core
aiohttp==3.10.0
python-telegram-bot==21.4
PyYAML==6.0.1
hydra-core==1.3.2
omegaconf==2.3.0
pydantic==2.8.2
tenacity==8.5.0

# Database
SQLAlchemy==2.0.31
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
pgvector==0.2.5

# RAG & AI Orchestration
langchain==0.2.12
langchain-community==0.2.10
langchain-core==0.2.29
langgraph==0.1.19
langsmith==0.1.96
langchain-text-splitters==0.2.2

# Search & Vector Stores
elasticsearch==8.14.0
langchain-elasticsearch==0.2.2
langchain-postgres==0.0.9

# Cloud LLM Adapters
langchain-openai==0.1.21
langchain-google-genai
google-genai
openai==1.40.3
tiktoken==0.7.0

# Document Parsing
beautifulsoup4==4.12.3
lxml==5.2.2
pypdf==4.3.1

# Utilities
attrs==23.2.0
certifi==2024.7.4
multidict==6.0.5
yarl==1.9.4
python-dateutil==2.9.0
tqdm==4.66.4
psutil==5.9.0
